requirements.txt
# downloading images from the web
requests>=2.31.0

# resizing images
pillow>=10.2.0

# for parsing svg files
svgelements>=1.9.6

# for progress bar
tqdm>=4.66.2

# dataset augmentation
numpy>=1.26.4
opencv-python>=4.9.0.80
scipy>=1.12.0
